rename BYOC package
