+++ src/pkg/cli/client/byoc/aws/stream.go
@@ -1,4 +1,4 @@
-package clouds
+package aws
 
 import (
 	"context"
